+++ src/ladb_opencutlist/ruby/constants.rb
@@ -1,7 +1,7 @@
 module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
-    EXTENSION_VERSION = '1.9.10'.freeze
-    EXTENSION_BUILD = '202012091141'.freeze
+    EXTENSION_VERSION = '2.0.0-dev'.freeze
+    EXTENSION_BUILD = '202101241601'.freeze
 
 end