+++ src/ladb_opencutlist/ruby/worker/cutlist/cutlist_cuttingdiagram_2d_worker.rb
@@ -6,8 +6,8 @@ module Ladb::OpenCutList
 
   class CutlistCuttingdiagram2dWorker
 
-    ORIGIN_POSITION_TOP_LEFT = 0
-    ORIGIN_POSITION_BOTTOM_LEFT = 1
+    ORIGIN_CORNER_TOP_LEFT = 0
+    ORIGIN_CORNER_BOTTOM_LEFT = 1
 
     def initialize(settings, cutlist)
       @group_id = settings['group_id']
@@ -23,7 +23,7 @@ module Ladb::OpenCutList
       @sheet_folding = settings['sheet_folding']
       @hide_cross = settings['hide_cross']
       @hide_part_list = settings['hide_part_list']
-      @origin_position = settings['origin_position'].to_i
+      @origin_corner = settings['origin_corner'].to_i
 
       @cutlist = cutlist
 
@@ -96,6 +96,7 @@ module Ladb::OpenCutList
               :trimming => options.trimsize.to_l.to_s,
               :optimization => @optimization,
               :stacking => @stacking,
+              :origin_corner => @origin_corner,
           },
 
           :unplaced_parts => [],
@@ -228,7 +229,7 @@ module Ladb::OpenCutList
                 :number => box.data.number,
                 :name => box.data.name,
                 :px_x => _to_px(box.x),
-                :px_y => _to_px(_compute_y_with_origin_position(@origin_position, box.y, box.width, bin.width)),
+                :px_y => _to_px(_compute_y_with_origin_corner(@origin_corner, box.y, box.width, bin.width)),
                 :px_length => _to_px(box.length),
                 :px_width => _to_px(box.width),
                 :length => box.data.cutting_length,
@@ -267,7 +268,7 @@ module Ladb::OpenCutList
             sheet[:leftovers].push(
                 {
                     :px_x => _to_px(box.x),
-                    :px_y => _to_px(_compute_y_with_origin_position(@origin_position, box.y, box.width, bin.width)),
+                    :px_y => _to_px(_compute_y_with_origin_corner(@origin_corner, box.y, box.width, bin.width)),
                     :px_length => _to_px(box.length),
                     :px_width => _to_px(box.width),
                     :length => box.length.to_l.to_s,
@@ -281,7 +282,7 @@ module Ladb::OpenCutList
             sheet[:cuts].push(
                 {
                     :px_x => _to_px(cut.x),
-                    :px_y => _to_px(_compute_y_with_origin_position(@origin_position, cut.y, cut.is_horizontal ? 0 : cut.length, bin.width)),
+                    :px_y => _to_px(_compute_y_with_origin_corner(@origin_corner, cut.y, cut.is_horizontal ? 0 : cut.length, bin.width)),
                     :px_length => _to_px(cut.length),
                     :x => cut.x.to_l.to_s,
                     :y => cut.y.to_l.to_s,
@@ -340,12 +341,12 @@ module Ladb::OpenCutList
       sheet[:total_area] += Size2d.new(bin.length.to_l, bin.width.to_l).area
     end
 
-    def _compute_y_with_origin_position(origin_position, y, y_size, y_translation)
-      case origin_position
-      when ORIGIN_POSITION_TOP_LEFT
-        y
-      when ORIGIN_POSITION_BOTTOM_LEFT
-        y_translation - y - y_size
+    def _compute_y_with_origin_corner(origin_corner, y, y_size, y_translation)
+      case origin_corner
+        when ORIGIN_CORNER_BOTTOM_LEFT
+          y_translation - y - y_size
+        else
+          y
       end
     end